Remove unnecessary type conversions in pkg/base
uint8(b[2]&0x0E), int(len(b)/2), and uint8(tlv.Value[0]) are all
redundant as the operands are already the target types.

+++ pkg/base/label.go
@@ -48,7 +48,7 @@ func MakeLabel(b []byte, srv6 ...bool) (*Label, error) {
 	}
 	l.Value >>= 4
 	// Move Exp bits to the beginning of the byte and leave only 3 bits, mask the rest.
-	l.Exp = uint8(b[2]&0x0E) >> 1
+	l.Exp = (b[2] & 0x0E) >> 1
 	l.BoS = b[2]&0x01 == 1
 
 	return &l, nil
